+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -19,6 +19,7 @@
 #include "Rule12p3Check.h"
 #include "Rule14p4Check.h"
 #include "Rule15p5Check.h"
+#include "Rule15p6Check.h"
 #include "Rule15p7Check.h"
 #include "Rule16p4Check.h"
 #include "Rule16p7Check.h"
@@ -55,6 +56,8 @@ public:
         "misrac-rule-14p4");
     CheckFactories.registerCheck<Rule15p5Check>(
         "misrac-rule-15p5");
+    CheckFactories.registerCheck<Rule15p6Check>(
+        "misrac-rule-15p6");
     CheckFactories.registerCheck<Rule15p7Check>(
         "misrac-rule-15p7");
     CheckFactories.registerCheck<Rule16p4Check>(