+++ clang-tools-extra/clang-tidy/misrac/Rule15p6Check.h
@@ -1,0 +1,38 @@
+//===--- Rule15p6Check.h - clang-tidy ---------------------------*- C++ -*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_MISRAC_RULE15P6CHECK_H
+#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_MISRAC_RULE15P6CHECK_H
+
+#include "MisracBaseCheck.h"
+
+namespace clang::tidy::misrac {
+
+/// Detects issues in accordance with MISRA C rule 15.6:
+/// "The body of an iteration-statement or a selection-statement shall be a
+/// compound-statement".
+///
+/// For the user-facing documentation see:
+/// http://clang.llvm.org/extra/clang-tidy/checks/misrac/rule-15p6.html
+class Rule15p6Check : public MisracBaseCheck {
+public:
+  Rule15p6Check(StringRef Name, ClangTidyContext *Context)
+      : MisracBaseCheck(Name, Context) {}
+
+  void addMatchers(ast_matchers::MatchFinder *Finder) override;
+
+protected:
+  void dispatch(const ast_matchers::MatchFinder::MatchResult &Result) override;
+
+private:
+  DiagnosticBuilder diag(const Stmt *Stmt);
+};
+
+} // namespace clang::tidy::misrac
+
+#endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_MISRAC_RULE15P6CHECK_H